+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/RerrorMessage.java
@@ -37,6 +37,9 @@ import uk.ac.rdg.resc.jstyx.StyxUtils;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -61,7 +64,7 @@ public class RerrorMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public RerrorMessage(int length, int type, int tag)
+    public RerrorMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Rerror";
@@ -69,7 +72,7 @@ public class RerrorMessage extends StyxMessage
     
     public RerrorMessage(String message)
     {
-        this(0, 107, 0); // The length and tag will be set when the message is sent
+        this(0, (short)107, 0); // The length and tag will be set when the message is sent
         this.setMessage(message);
     }
