+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/RwriteMessage.java
@@ -35,6 +35,9 @@ package uk.ac.rdg.resc.jstyx.messages;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -59,7 +62,7 @@ public class RwriteMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public RwriteMessage(int length, int type, int tag)
+    public RwriteMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Rwrite";
@@ -67,7 +70,7 @@ public class RwriteMessage extends StyxMessage
     
     public RwriteMessage()
     {
-        this(11, 119, 0); // The tag will be added later
+        this(11, (short)119, 0); // The tag will be added later
     }
     
     public RwriteMessage(long count)
