+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TattachMessage.java
@@ -37,6 +37,9 @@ import uk.ac.rdg.resc.jstyx.StyxUtils;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -67,7 +70,7 @@ public class TattachMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TattachMessage(int length, int type, int tag)
+    public TattachMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Tattach";
@@ -75,7 +78,7 @@ public class TattachMessage extends StyxMessage
     
     public TattachMessage(long fid, long afid, String uname, String aname)
     {
-        this(0, 104, 0); // The length and tag will be added later
+        this(0, (short)104, 0); // The length and tag will be added later
         this.fid = fid;
         this.afid = afid;
         this.uname = uname;
