+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TclunkMessage.java
@@ -35,6 +35,9 @@ package uk.ac.rdg.resc.jstyx.messages;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -60,7 +63,7 @@ public class TclunkMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TclunkMessage(int length, int type, int tag)
+    public TclunkMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Tclunk";
@@ -68,7 +71,7 @@ public class TclunkMessage extends StyxMessage
     
     public TclunkMessage(long fid)
     {
-        this(11, 120, 0); // The tag will be added later
+        this(11, (short)120, 0); // The tag will be added later
         this.fid = fid;
     }
