+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TflushMessage.java
@@ -36,6 +36,9 @@ package uk.ac.rdg.resc.jstyx.messages;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:15  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -61,7 +64,7 @@ public class TflushMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TflushMessage(int length, int type, int tag)
+    public TflushMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Tflush";
@@ -72,7 +75,7 @@ public class TflushMessage extends StyxMessage
      */
     public TflushMessage(int oldTag)
     {
-        this(9, 108, 0); // The tag will be set later
+        this(9, (short)108, 0); // The tag will be set later
         this.oldTag = oldTag;
     }
