+++ core/src/java/uk/ac/rdg/resc/jstyx/messages/TversionMessage.java
@@ -38,6 +38,9 @@ import uk.ac.rdg.resc.jstyx.StyxUtils;
  * $Revision$
  * $Date$
  * $Log$
+ * Revision 1.4  2005/03/15 09:01:48  jonblower
+ * Message type now stored as short, not int
+ *
  * Revision 1.3  2005/03/11 14:02:16  jonblower
  * Merged MINA-Test_20059309 into main line of development
  *
@@ -68,7 +71,7 @@ public class TversionMessage extends StyxMessage
      * @param type The type of the message (a number between 100 and 127)
      * @param tag The tag that identifies this message
      */
-    public TversionMessage(int length, int type, int tag)
+    public TversionMessage(int length, short type, int tag)
     {
         super(length, type, tag);
         this.name = "Tversion";
@@ -80,7 +83,7 @@ public class TversionMessage extends StyxMessage
      */
     public TversionMessage()
     {
-        this(19, 100, StyxUtils.NOTAG);
+        this(19, (short)100, StyxUtils.NOTAG);
         this.maxMessageSize = 8216;
         this.setVersion("9P2000");
     }
